+++ ICD.Connect.Sources.Barco/Devices/ClickshareConference/BarcoClickshareConferenceDevice.cs
@@ -6,14 +6,18 @@ using ICD.Common.Utils.Collections;
 using ICD.Common.Utils.EventArguments;
 using ICD.Common.Utils.Extensions;
 using ICD.Common.Utils.Services.Logging;
+using ICD.Connect.API.Nodes;
+using ICD.Connect.Conferencing.BYOD;
+using ICD.Connect.Devices;
 using ICD.Connect.Devices.Controls;
+using ICD.Connect.Devices.Extensions;
 using ICD.Connect.Settings;
 using ICD.Connect.Sources.Barco.API;
 using ICD.Connect.Sources.Barco.Responses.v2;
 
 namespace ICD.Connect.Sources.Barco.Devices.ClickshareConference
 {
-	public sealed class BarcoClickshareConferenceDevice : AbstractBarcoClickshareDevice<BarcoClickshareConferenceDeviceSettings>
+	public sealed class BarcoClickshareConferenceDevice : AbstractBarcoClickshareDevice<BarcoClickshareConferenceDeviceSettings>, IByodHubDevice
 	{
 		#region Events
 
@@ -41,6 +45,11 @@ namespace ICD.Connect.Sources.Barco.Devices.ClickshareConference
 				return apiV2;
 			}
 		}
+
+		/// <summary>
+		/// Destination device for cameras routed through a BYOD Hub Device.
+		/// </summary>
+		public IDevice CameraDestinationDevice { get; set; }
 
 		#endregion
 
@@ -124,6 +133,52 @@ namespace ICD.Connect.Sources.Barco.Devices.ClickshareConference
 			base.AddControls(settings, factory, addControl);
 
 			addControl(new BarcoClickshareConferenceDeviceConferenceControl(this, 1));
+		}
+
+		#endregion
+
+		#region Settings
+
+		protected override void ClearSettingsFinal()
+		{
+			base.ClearSettingsFinal();
+
+			CameraDestinationDevice = null;
+		}
+
+		protected override void CopySettingsFinal(BarcoClickshareConferenceDeviceSettings settings)
+		{
+			base.CopySettingsFinal(settings);
+
+			settings.CameraDestinationDevice = CameraDestinationDevice == null ? (int?)null : CameraDestinationDevice.Id;
+		}
+
+		protected override void ApplySettingsFinal(BarcoClickshareConferenceDeviceSettings settings, IDeviceFactory factory)
+		{
+			base.ApplySettingsFinal(settings, factory);
+
+			if (settings.CameraDestinationDevice == null)
+				return;
+
+			try
+			{
+				CameraDestinationDevice = factory.GetDeviceById(settings.CameraDestinationDevice.Value);
+			}
+			catch (KeyNotFoundException)
+			{
+				Logger.Log(eSeverity.Error, "No device with id {0}", settings.CameraDestinationDevice);
+			}
+		}
+
+		#endregion
+
+		#region Console
+
+		public override void BuildConsoleStatus(AddStatusRowDelegate addRow)
+		{
+			base.BuildConsoleStatus(addRow);
+
+			addRow("CameraDestinationDevice", CameraDestinationDevice);
 		}
 
 		#endregion