+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
  - Barco Clickshare - Clarify API Version in console status
  - Remove StreamTvTunerRouteSourceControl & have StreamTvTunerDevice use new StreamSourceDeviceRoutingControl
+ - BarcoClickshareConferenceDevice now implements IByodHubDevice
 
 ## [7.1.0] - 2021-01-14
 ### Added